refactor() remove legacy code and some renaming
