+++ src/core/loading/service.py
@@ -10,7 +10,6 @@ from ...setup.config import PathConfig
 from ...setup.logging import logger
 from ...database.schemas import Database
 from ..schemas import AuditMetadata
-
 from .strategies import BaseDataLoadingStrategy
